updated snippets list to navigate to snippet page
